Account for rootfs compression and total size
A heuristic was used in the previous implementation for
simplicity (*2) but when using bigger file systems size
overflows.

This patch uncompresses the archive in an auxiliary
folder and calculates its total size in bytes instead.
In addition, the slack has been extended to 512 MB so
that there's enough size for experimentation with the
images.

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>